Fix settings not working
The GUI got a different instance of the Settings class, meaning it
couldn't see the settings actually set/loaded.

This moves Settings into its own class and moves many other things
around to make that work.
