+++ pext/appfile.py
@@ -1,0 +1,34 @@
+#!/usr/bin/env python3
+
+# Copyright (c) 2015 - 2022 Sylvia van Os <sylvia@hackerchick.me>
+#
+# This file is part of Pext.
+#
+# Pext is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Pext.
+
+This is Pext's AppFile class.
+"""
+
+import os
+
+
+class AppFile():
+    """Get access to application-specific files."""
+
+    @staticmethod
+    def get_path() -> str:
+        """Return the absolute current path."""
+        return os.path.dirname(os.path.abspath(__file__))
